+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-07T08:05:54.516Z",
+    "2026-05-07T12:33:45.946Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {