+++ packages/orq-rc/RELEASES.md
@@ -8919,3 +8919,13 @@ Based on:
 - [typescript v4.9.0-rc.30] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.30 - packages/orq-rc
+
+## 2026-05-07 12:37:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.31] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.31] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.31 - packages/orq-rc